Update index.html
At the moment I have made the link be shortened for the data base. The database can only be seen and the user can't edit it. At the moment I am going to split the database in order to create different types of tiers. In order to do that, I need to make different copies of the code in order to have the links of the database be able to "seen" through the different links.