Create tier1v3.html
I decided that I am going with the new display as it keeps the user from copying and essentially stealing any of the information that we have gathered.  This is the third of the first tier. So now I am going to move on to the second tier and so.